# backend/docs/swagger.yaml
basePath: /
definitions:
...
  models.PointRequest:
    properties:
      end_at:
        description: Время окончания события
        example: "2025-12-01T10:00:00Z"
        type: string
      event_id:
        description: ID мероприятия, к которому относится точка (обязательно)
        example: a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11
        type: string
      id:
        description: ID точки (обязательно для обновления)
        example: 11eebc99-9c0b-4ef8-bb6d-6bb9bd380a11
        type: string
      start_at:
        description: Время начала события
        example: "2025-12-01T09:00:00Z"
        type: string
      timeline_description:
        description: |-
          Поля Таймлайна
          Описание события в таймлайне
        example: Регистрация участников и утренний кофе.
        type: string
      title:
        description: Заголовок точки (обязательно)
        example: Стойка регистрации
        type: string
      x:
        description: Координата X на карте (обязательно)
        example: 50
        type: integer
      "y":
        description: Координата Y на карте (обязательно)
        example: 100
        type: integer
    type: object
  models.PointResponse:
    properties:
      end_at:
        description: Время окончания события
        example: "2025-12-01T10:00:00Z"
        type: string
      event_id:
        description: ID мероприятия, к которому относится точка
        example: a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11
        type: string
      id:
        description: ID точки
        example: 11eebc99-9c0b-4ef8-bb6d-6bb9bd380a11
        type: string
      start_at:
        description: Время начала события
        example: "2025-12-01T09:00:00Z"
        type: string
      timeline_description:
        description: |-
          Поля Таймлайна (поля могут быть NULLable в базе, поэтому используем указатели, если они не являются обязательными)
          Описание события в таймлайне
        example: Регистрация участников и утренний кофе.
        type: string
      title:
        description: Заголовок точки
        example: Стойка регистрации
        type: string
      x:
        description: Координата X на карте
        example: 50
        type: integer
      "y":
        description: Координата Y на карте
        example: 100
        type: integer
    type: object
...
  models.SimplePointRequest:
    properties:
      event_id:
        description: ID мероприятия, к которому относится точка (обязательно)
        example: a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11
        type: string
      id:
        description: ID точки (обязательно для обновления)
        example: 44eebc99-9c0b-4ef8-bb6d-6bb9bd380a11
        type: string
      title:
        description: Заголовок точки (обязательно)
        example: Название
        type: string
      x:
        description: Координата X на карте (обязательно)
        example: 90
        type: integer
      "y":
        description: Координата Y на карте (обязательно)
        example: 150
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Сервис для управления событиями, картами и точками интереса.
  title: Event Designer API
  version: "1.0"
paths:
...
  /admin/events/{eventID}/points/{pointID}:
    delete:
      description: Удаляет простую точку или точку таймлайна по ее ID.
      parameters:
      - description: ID мероприятия (UUID)
        in: path
        name: eventID
        required: true
        type: string
      - description: ID точки (UUID)
        in: path
        name: pointID
        required: true
        type: string
      responses:
        "204":
          description: Точка успешно удалена (No Content)
        "400":
          description: Invalid ID format in URL
          schema:
            type: string
        "500":
          description: Failed to delete point
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Удалить точку
      tags:
      - Points
...
  /admin/events/points/simple:
    patch:
      consumes:
      - application/json
      description: Обновляет координаты и/или заголовок существующей простой точки.
      parameters:
      - description: ID точки и поля для обновления (ID required)
        in: body
        name: point
        required: true
        schema:
          $ref: '#/definitions/models.SimplePointRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Точка успешно обновлена
        "400":
          description: Invalid request body / Point ID must be provided
          schema:
            type: string
        "403":
          description: 'Forbidden: Cannot update Timeline Point using Simple Point
            endpoint / Cannot add timeline data'
          schema:
            type: string
        "404":
          description: Point not found
          schema:
            type: string
        "500":
          description: Failed to update simple point
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Обновить простую точку
      tags:
      - Points
    post:
      consumes:
      - application/json
      description: Создает новую точку на карте (без привязки к таймлайну). Требуется
        EventID, X, Y и Title.
      parameters:
      - description: Данные точки (X, Y, Title, EventID)
        in: body
        name: point
        required: true
        schema:
          $ref: '#/definitions/models.SimplePointRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Успешное создание, возвращает ID
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body / Required fields missing
          schema:
            type: string
        "403":
          description: 'Forbidden: Cannot use timeline data'
          schema:
            type: string
        "500":
          description: Failed to create simple point
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Создать простую точку
      tags:
      - Points
  /admin/events/points/timeline:
    patch:
      consumes:
      - application/json
      description: Обновляет существующую точку таймлайна. Требуется предоставление
        всех полей таймлайна для обновления.
      parameters:
      - description: ID точки и полные данные таймлайна (ID, StartAt, EndAt, Description
          required)
        in: body
        name: point
        required: true
        schema:
          $ref: '#/definitions/models.PointRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Точка успешно обновлена
        "400":
          description: Invalid request body / Required fields missing
          schema:
            type: string
        "403":
          description: 'Forbidden: Cannot update Simple Point using Timeline Point
            endpoint'
          schema:
            type: string
        "404":
          description: Point not found
          schema:
            type: string
        "500":
          description: Failed to update timeline point
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Обновить точку таймлайна
      tags:
      - Points
    post:
      consumes:
      - application/json
      description: Создает новую точку на карте с привязкой к таймлайну (StartAt,
        EndAt, Description). Требуются все поля.
      parameters:
      - description: Данные точки (EventID, X, Y, Title, StartAt, EndAt, TimelineDescription)
        in: body
        name: point
        required: true
        schema:
          $ref: '#/definitions/models.PointRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Успешное создание, возвращает ID
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body / All timeline fields are required
          schema:
            type: string
        "500":
          description: Failed to create timeline point
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Создать точку таймлайна
      tags:
      - Points
...
  /events/{eventID}/points:
    get:
      description: Возвращает список всех простых точек и точек таймлайна для указанного
        мероприятия.
      parameters:
      - description: ID мероприятия (UUID)
        in: path
        name: eventID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список точек
          schema:
            items:
              $ref: '#/definitions/models.PointResponse'
            type: array
        "400":
          description: Invalid event ID
          schema:
            type: string
        "500":
          description: Failed to fetch points
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Получить все точки по ID мероприятия
      tags:
      - Points
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"